fontend of properties list in profile finished, not yet connected to endpoint
