+++ src/pages/Profile.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
+import ProductListBody from '../components/products/ProductListBody';
 
 function ProfilePage() {
     const accessToken = localStorage.getItem('accessToken');
@@ -8,6 +9,7 @@ function ProfilePage() {
 
     const [updateStatus, setUpdateStatus] = useState('');
     const [userData, setUserData] = useState();
+    const [userProperties, setUserProperties] = useState();
 
     const updateUserFields = (e) => {
         const { name, value } = e.target;
@@ -35,6 +37,16 @@ function ProfilePage() {
             }
         }
         getUser();
+
+        async function getUserProperties() {
+            try {
+                let response = await axios.get(`http://localhost:8005/properties/${userData.username}`, { params: { accessToken: accessToken } });
+                setUserProperties(response.data);
+            } catch (error) {
+                console.error('Error fetching user properties:', error);
+            }
+        }
+        getUserProperties();
     }, []);
 
     if (userData) {
@@ -42,8 +54,8 @@ function ProfilePage() {
             <div className="container mt-5">
                 {/* Profile widget */}
                 <div className="row gutters-sm">
-                    <div className="col-md-4 mb-3">
-                        <div className="card">
+                    <div className="col-md-4 mb-3 flex-fill">
+                        <div className="card h-100">
                             <div className="card-body">
                                 <div className="d-flex flex-column align-items-center text-center">
                                     <div className="mt-3">
@@ -60,6 +72,7 @@ function ProfilePage() {
                     </div>
                     <div className="col-md-8">
                         <div className="card mb-3">
+                            <h4 className="card-header">Personal Details</h4>
                             <div className="card-body">
                                 {updateStatus === 'success' && <div className="alert alert-success">Update successful!</div>}
                                 {updateStatus === 'error' && <div className="alert alert-danger">Update failed. Please try again.</div>}
@@ -69,6 +82,14 @@ function ProfilePage() {
                                     <ProfileField title="Phone Number" type="tel" name="phoneNumber" value={userData.phoneNumber} onChange={updateUserFields} />
                                     <button type="submit" className="btn btn-primary">Update</button>
                                 </form>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="col-md-12">
+                        <div className="card mb-3">
+                            <h4 className="card-header">Your Properties</h4>
+                            <div className="card-body">
+                                <ProductListBody products={userProperties} />
                             </div>
                         </div>
                     </div>
