+++ src/products/ProductList.js
@@ -1,11 +1,9 @@
 import { Link, useLocation } from "react-router-dom";
-import Product from "./Product";
-import ProductH from "./ProductH";
-import { useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import ScrollToTopOnMount from "../template/ScrollToTopOnMount";
 import { housesData } from "../mockdata";
 import { useSearch } from '../SearchContext';
+import ProductListBody from "../components/products/ProductListBody";
 
 const categories = [
   "Houses for Sale",
@@ -71,25 +69,25 @@ function FilterMenuLeft() {
         </div>
       </li>
       <li className="list-group-item">
-      <h5 className="mt-1 mb-1">Location</h5>
-      <div className="col-lg-3 d-none d-lg-block">
-                <select
-                  className="form-select"
-                  aria-label="Default select example"
-                  defaultValue=""
-                >
-                  <option value="">Zurich</option>
-                  <option value="1">Basel</option>
-                  <option value="2">Geneva</option>
-                  <option value="3">Lausanne</option>
-                  <option value="4">Bern</option>
-                  <option value="4">Lucerne</option>
-                  <option value="4">St. Gallen</option>
-                  <option value="4">Biels</option>
-                  <option value="4">Thun</option>
-                  <option value="4">Winterthur</option>
-                </select>
-              </div>
+        <h5 className="mt-1 mb-1">Location</h5>
+        <div className="col-lg-3 d-none d-lg-block">
+          <select
+            className="form-select"
+            aria-label="Default select example"
+            defaultValue=""
+          >
+            <option value="">Zurich</option>
+            <option value="1">Basel</option>
+            <option value="2">Geneva</option>
+            <option value="3">Lausanne</option>
+            <option value="4">Bern</option>
+            <option value="4">Lucerne</option>
+            <option value="4">St. Gallen</option>
+            <option value="4">Biels</option>
+            <option value="4">Thun</option>
+            <option value="4">Winterthur</option>
+          </select>
+        </div>
       </li>
       <li className="list-group-item">
         <h5 className="mt-1 mb-2">Price Range</h5>
@@ -120,19 +118,12 @@ function FilterMenuLeft() {
 }
 
 function ProductList() {
-  const [viewType, setViewType] = useState({ grid: true });
   //const results = housesData;
   const location = useLocation();
   const { searchResults } = useSearch();
- 
-  const results = searchResults || housesData;
-  //const results = searchResults.length > 0 ? searchResults : housesData;
 
-  function changeViewType() {
-    setViewType({
-      grid: !viewType.grid,
-    });
-  }
+  // const results = searchResults || housesData;
+  const results = searchResults.length > 0 ? searchResults : housesData;
 
   return (
     <div className="container mt-5 py-4 px-xl-5">
@@ -201,7 +192,6 @@ function ProductList() {
           </div>
         </div>
       </div>
-
       <div className="row mb-4 mt-lg-3">
         <div className="d-none d-lg-block col-lg-3">
           <div className="border rounded shadow-sm">
@@ -209,96 +199,7 @@ function ProductList() {
           </div>
         </div>
         <div className="col-lg-9">
-          <div className="d-flex flex-column h-100">
-            <div className="row mb-3">
-              <div className="col-lg-3 d-none d-lg-block">
-                <select
-                  className="form-select"
-                  aria-label="Default select example"
-                  defaultValue=""
-                >
-                  <option value="">Newest first</option>
-                  <option value="1">Price (lowest first)</option>
-                  <option value="2">Price (highest first)</option>
-                  <option value="3">Surface (highest first)</option>
-                  <option value="4">Surface (lowest first)</option>
-                </select>
-              </div>
-              <div className="col-lg-9 col-xl-5 offset-xl-4 d-flex flex-row">
-                <div className="input-group">
-                  <input
-                    className="form-control"
-                    type="text"
-                    placeholder="Search products..."
-                    aria-label="search input"
-                  />
-                  <button className="btn btn-outline-dark">
-                    <FontAwesomeIcon icon={["fas", "search"]} />
-                  </button>
-                </div>
-                <button
-                  className="btn btn-outline-dark ms-2 d-none d-lg-inline"
-                  onClick={changeViewType}
-                >
-                  <FontAwesomeIcon
-                    icon={["fas", viewType.grid ? "th-list" : "th-large"]}
-                  />
-                </button>
-              </div>
-            </div>
-            <div
-          className={
-            "row row-cols-1 row-cols-md-2 row-cols-lg-2 g-3 mb-4 flex-shrink-0 " +
-            (viewType.grid ? "row-cols-xl-3" : "row-cols-xl-2")
-          }
-        >
-          {results.map((product, i) => {
-            if (viewType.grid) {
-              return (
-                <Product key={product.propertyId} product={product} />
-              );
-            }
-            return (
-              <ProductH key={product.propertyId} product={product} />
-            );
-          })}
-        </div>
-
-            <div className="d-flex align-items-center mt-auto">
-              <span className="text-muted small d-none d-md-inline">
-                Showing 10 of 100
-              </span>
-              <nav aria-label="Page navigation example" className="ms-auto">
-                <ul className="pagination my-0">
-                  <li className="page-item">
-                    <a className="page-link" href="!#">
-                      Previous
-                    </a>
-                  </li>
-                  <li className="page-item">
-                    <a className="page-link" href="!#">
-                      1
-                    </a>
-                  </li>
-                  <li className="page-item active">
-                    <a className="page-link" href="!#">
-                      2
-                    </a>
-                  </li>
-                  <li className="page-item">
-                    <a className="page-link" href="!#">
-                      3
-                    </a>
-                  </li>
-                  <li className="page-item">
-                    <a className="page-link" href="!#">
-                      Next
-                    </a>
-                  </li>
-                </ul>
-              </nav>
-            </div>
-          </div>
+          <ProductListBody products={results} />
         </div>
       </div>
     </div>
